Converts AdminsController test partially to rspec, needs work
